update: convert the composable to a more generic graphql consumer and take aboard feedback
